feat(frontent): re-adds onboarding wizard
